+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Implemented `iscss` function to identify whether a given code is known to be a CSS (Calderbank-Shor-Steane) code.
 - Implemented the classical Reed-Muller code
 - Added the surface code to the ECC module.
+- Added the Steane-Reed-Muller code to the ECC module.
  
 ## v0.9.0 - 2024-03-19